src/requirements.txt
transformers==4.44.2
pytorch-lightning==2.1.2
pandas==2.1.3
python-dotenv==0.21.0
gitpython==3.1.31
wandb==0.17.1
ipykernel==6.29.5